+++ dune/localfunctions/lagrange/pk1d/pk1dlocalinterpolation.hh
@@ -4,6 +4,7 @@
 #define DUNE_Pk1DLOCALINTERPOLATION_HH
 
 #include <vector>
+#include <dune/localfunctions/common/localinterpolation.hh>
 
 namespace Dune
 {
@@ -22,18 +23,19 @@ namespace Dune
   public:
 
     template<typename F, typename C>
-    void interpolate (const F& f, std::vector<C>& out) const
+    void interpolate (const F& ff, std::vector<C>& out) const
     {
       typename LB::Traits::DomainType x;
-      typename LB::Traits::RangeType y;
       typedef typename LB::Traits::DomainFieldType D;
+
+      auto&& f = Impl::makeFunctionWithCallOperator<decltype(x)>(ff);
+
       out.resize(N);
 
       for (int i=0; i<N; i++)
       {
         x[0] = ((D)i)/((D)kdiv);
-        f.evaluate(x,y);
-        out[i] = y;
+        out[i] = f(x);
       }
     }