+++ dune/localfunctions/lagrange/pk3d/pk3dlocalinterpolation.hh
@@ -18,11 +18,13 @@ namespace Dune
   public:
 
     template<typename F, typename C>
-    void interpolate (const F& f, std::vector<C>& out) const
+    void interpolate (const F& ff, std::vector<C>& out) const
     {
       typename LB::Traits::DomainType x;
-      typename LB::Traits::RangeType y;
       typedef typename LB::Traits::DomainFieldType D;
+
+      auto&& f = Impl::makeFunctionWithCallOperator<typename LB::Traits::DomainType>(ff);
+
       out.resize(N);
       int n=0;
       for (int i2 = 0; i2 <= k; i2++)
@@ -32,8 +34,7 @@ namespace Dune
             x[0] = ((D)i0)/((D)kdiv);
             x[1] = ((D)i1)/((D)kdiv);
             x[2] = ((D)i2)/((D)kdiv);
-            f.evaluate(x,y);
-            out[n] = y;
+            out[n] = f(x);
             n++;
           }
     }