+++ dune/localfunctions/whitney/edges0.5/interpolation.hh
@@ -8,6 +8,7 @@
 #include <vector>
 
 #include <dune/localfunctions/whitney/edges0.5/common.hh>
+#include <dune/localfunctions/common/localinterpolation.hh>
 
 namespace Dune {
 
@@ -69,13 +70,15 @@ namespace Dune {
 
     //! Interpolation of a function
     template<typename F, typename C>
-    void interpolate(const F& f, std::vector<C>& out) const {
+    void interpolate(const F& ff, std::vector<C>& out) const {
       typename Traits::Range y;
+
+      auto&& f = Impl::makeFunctionWithCallOperator<std::decay_t<decltype(refelem.position(0,dim-1))>>(ff);
 
       out.resize(s);
 
       for(std::size_t i = 0; i < s; ++i) {
-        f.evaluate(refelem.position(i,dim-1), y);
+        y = f(refelem.position(i,dim-1));
 
         out[i] = y * edgev[i];
       }